bot/requirements.txt
discord.py

# DO NOT REMOVE THIS, THESE ARE THE REQUIREMENTS FOR THE BOT TO WORK. 